python_scripts/requirements.txt
cython
pyjnius